+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+- Fix: Default OIDC provider name not set while updating to Nextcloud Hub setup [#1130](https://github.com/nextcloud/integration_openproject/pull/1130)
+
 ### Changed
 
 ### Removed