+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Support Nextcloud 34 [#1021](https://github.com/nextcloud/integration_openproject/pull/1021)
+
 ### Changed
 
 ### Removed