+++ CHANGELOG.md
@@ -13,17 +13,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
 ### Fixed
 
-- Fix: Wrong message "Unauthorized to connect to OpenProject" shown with valid SSO connection [#1018](https://github.com/nextcloud/integration_openproject/pull/1018)
+### Changed
+
+### Removed
+
+## 3.1.0 - 2026-06-18
 
 ### Added
 
 - Support Nextcloud 34 [#1021](https://github.com/nextcloud/integration_openproject/pull/1021)
 
-### Changed
+### Fixed
 
-### Removed
+- Fix: Wrong message "Unauthorized to connect to OpenProject" shown with valid SSO connection [#1018](https://github.com/nextcloud/integration_openproject/pull/1018)
+- Fix: No OpenProject notification in Nextcloud dashboard [#1045](https://github.com/nextcloud/integration_openproject/pull/1045)
 
 ## 3.0.0 - 2026-03-17